- Organize the file structure - Fix: ESLint errors
