+++ .prettierrc.json
@@ -1,0 +1,6 @@
+{
+  "bracketSpacing": true,
+  "singleQuote": true,
+  "trailingComma": "es5",
+  "arrowParens": "avoid"
+}
