+++ src/domain/calculator/index.ts
@@ -1,0 +1,3 @@
+export * from './custom-collection';
+export * from './icalculator';
+export * from './simple-calculator';
