+++ src/domain/index.ts
@@ -1,0 +1,2 @@
+export * from './calculator';
+export * from './model';
