+++ src/domain/model/index.ts
@@ -1,0 +1,3 @@
+export * from './action-value';
+export * from './number-value';
+export * from './value-kind';
