+++ src/index.css
@@ -1,4 +1,4 @@
-@charset "UTF-8";
+@charset 'UTF-8';
 
 body {
   position: absolute;
