+++ src/interface-adapter/main-view/index.ts
@@ -1,0 +1,2 @@
+export * from './mainview-databind';
+export * from './mainview-viewmodel';
