+++ src/main.ts
@@ -1,22 +1,16 @@
-import { Application } from "./interface-adapter/application/application";
-import { DataBind } from "./interface-adapter/ui/data-bind.class";
-import { CustomCollection } from "./infrastructure/custom-collection";
-import { IButtonValue } from "./interface-adapter/ui/pad-button";
-import { SimpleCalculator } from "./domain/calculator/simple-calculator";
+import { CustomCollection, SimpleCalculator } from './domain';
+import { MainViewDataBind, MainViewViewModel } from './interface-adapter';
+import { IButtonValue } from './usecase';
 
-const dataBind = new DataBind();
+const dataBind = new MainViewDataBind();
 const equation = new CustomCollection<IButtonValue>();
 const calculator = new SimpleCalculator();
-const app = new Application(
-    dataBind,
-    equation,
-    calculator,
-);
+const app = new MainViewViewModel(dataBind, equation, calculator);
 
-window.addEventListener("load", () => {
-    app.start();
+window.addEventListener('load', () => {
+  app.start();
 });
 
-window.addEventListener("unload", () => {
-    app.destroy();
+window.addEventListener('unload', () => {
+  app.destroy();
 });
